for guest users: add to cart, remove from cart, cart persists in store, cart persists in local storage

+++ client/components/CartSubtotal.js
@@ -1,21 +1,33 @@
-import React, { useState, useEffect } from 'react';
+import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { getCartThunk, deleteItemThunk } from '../store/cart';
+import { getCartThunk, deleteItemThunk, getCart } from '../store/cart';
 
 // grab cart
 const CartSubtotal = () => {
   const dispatch = useDispatch();
-  const { orderId, cart } = useSelector((state) => state);
-
+  let { orderId, cart, auth } = useSelector((state) => state);
   useEffect(() => {
-    dispatch(getCartThunk(orderId));
+    if (auth.id) {
+      dispatch(getCartThunk(orderId));
+    } else {
+      cart = JSON.parse(window.localStorage.getItem('cart'));
+      dispatch(getCart(cart));
+    }
   }, [orderId]);
 
   const handleDelete = async (productId, orderId) => {
-    await dispatch(deleteItemThunk(productId, orderId));
-    dispatch(getCartThunk(orderId));
+    if (auth.id) {
+      await dispatch(deleteItemThunk(productId, orderId));
+      dispatch(getCartThunk(orderId));
+    } else {
+      cart = JSON.parse(window.localStorage.getItem('cart'));
+      cart = cart.filter((item) => item.productId !== productId);
+      window.localStorage.setItem('cart', JSON.stringify(cart));
+      dispatch(getCart(cart));
+    }
   };
 
+  console.log(cart);
   return (
     <>
       <div>CartSubtotal</div>
